Very rough implementation of <font> tag
Adds an additional option to the `options` object, called
`createFontTags`. If the option is set to "true", then while looping
through the document elements to create style props, changes the
behavior of the `setStyleAttrs` method:

- if the element has no children, and the style attribute contains any
  of `font-family`, `font-size`, or `color`, those elements are not
  added to the style attribute, but instead are added to a <font> tag
  which is wrapped around the element's innerHTML.

This needs a LOT of sanity checking and test cases before being anything
like ready for production... just putting it out to see if it even makes
sense to have in here.